pyproject.toml
[project]
name = "mailhub"
version = "0.3.1"
description = "Unified email/calendar/contacts connector for OpenClaw skill MailHub"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12",
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "rich>=13.7",
  "keyring>=25.2",
  "cryptography>=43.0",
  "requests>=2.32",
  "httpx>=0.27",
  "python-dateutil>=2.9",
  "sqlite-utils>=3.36",
  "beautifulsoup4>=4.12",
  "lxml>=5.2",
  "pdfplumber>=0.11",
]

[project.scripts]
mailhub = "mailhub.cli:app"

[tool.uv]
package = true